==> 1. ProjectPrediksiGaji_DataCollection.ipynb <==
#1. Salary Prediction Project - Data Collection Part 1
by: Cindy Suyitno

The data will be collected from:

https://glints.com/id/opportunities/jobs/explore?country=ID&locationName=Indonesia&gclid=Cj0KCQjwkruVBhCHARIsACVIiOwkqvkaoBKtrxO2n5JAlucbZD1OCfdhStJ7iMXIChg6ou--n6m-xAkaAj8DEALw_wcB&utm_source=google&utm_medium=cpc&utm_campaign=ID|Marketplace|Search|Brand&gclid=Cj0KCQjwkruVBhCHARIsACVIiOwkqvkaoBKtrxO2n5JAlucbZD1OCfdhStJ7iMXIChg6ou--n6m-xAkaAj8DEALw_wcB 

aka Glints webpage, one of the famous website with Job vacancies for Indonesian.

File that is extracted from this part:
- DataCollection_Glints.csv (--> will be used in part 3. Data Cleaning)

#There are some modules need to be install for this data collection process
!pip install beautifulsoup4
!pip install lxml
!pip install selenium
!pip install CurrencyConverter

#since this file is used in googlecollab, there are some prior setting as well for the chromedriver, which is esential in this step
!apt-get update
!apt install chromium-chromedriver

#this syntax will mount my google colab file to my google drive
from google.colab import drive
drive.mount ('/content/drive/', force_remount=True)

%cd /content/drive/MyDrive/Colab Notebooks 
#change directory

#library imports
from bs4 import BeautifulSoup
import requests
import lxml
import pandas as pd
import numpy as np
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
import os
from random import randint
from time import sleep
from currency_converter import CurrencyConverter

#these bunch of codes are important to establish webdriver.Chrome, which is a python tool for web scraping
!cp /usr/lib/chromium-browser/chromedriver /usr/bin
import sys
sys.path.insert(0,'/usr/lib/chromium-browser/chromedriver')

chrome_options = webdriver.ChromeOptions()
chrome_options.add_argument('--headless')
chrome_options.add_argument('--no-sandbox')
chrome_options.add_argument('--disable-dev-shm-usage')
wd = webdriver.Chrome('chromedriver',chrome_options=chrome_options)

#this is the data href i scrapped on 6/9/2022, on glints.com (which has 300 pages)
glints_site = []
for a in range(300):
  glints_site.append('https://glints.com/id/opportunities/jobs/explore?country=ID&locationName=Indonesia&gclid=CjwKCAjwkYGVBhArEiwA4sZLuO7UuluK-tMrs4ZCZp-E7ZMeembyo3lSz9hwCQofjoeLs0MBqVmsNBoC394QAvD_BwE&gclid=CjwKCAjwkYGVBhArEiwA4sZLuO7UuluK-tMrs4ZCZp-E7ZMeembyo3lSz9hwCQofjoeLs0MBqVmsNBoC394QAvD_BwE&utm_source=google&utm_medium=cpc&utm_campaign=ID%7CMarketplace%7CSearch%7CBrand&page='+str(a+1))

#there are 4 items i want to scrap from this website: job name, city, salary, and years of experiences
#these are the class name of those items (can be retrieved from click right -> Inspect to the text in the website)
#job name = CompactOpportunityCardsc__JobTitle-sc-1y4v110-7 bVuGlP
#kota, gaji, yrs of experience = CompactOpportunityCardsc__OpportunityInfo-sc-1y4v110-13 ikxvyY

my_table_data1_1 = []
my_table_data1_2 = []
for i in range(0,300):
  url = glints_site[i]
  wd.get(url)
  sleep(randint(1,3)) #slow down process to avoid error
  soup = BeautifulSoup(wd.page_source, 'html.parser')
  my_table = soup.find_all(class_=['CompactOpportunityCardsc__JobTitle-sc-1y4v110-7 bVuGlP'])
  for data in my_table:
      my_table_data1_1.append(data.get_text())
  my_table2 = soup.find_all(class_=['CompactOpportunityCardsc__OpportunityInfo-sc-1y4v110-13 ikxvyY'])
  for data in my_table2:
      my_table_data1_2.append(data.get_text()) 

#assigning each item to salary, city, and years
cc = CurrencyConverter()
currency = list(cc.currencies)

def is_this_salary(a):
  for y in range(len(currency)):
    if currency[y] in str(a):
      return True
    elif 'salary' in str(a):
      return True
  else:
    return False

salary = []
city = []
years = []
x=0

try:
  for x in range(len(my_table_data1_2)): 
    if ('tahun' not in my_table_data1_2[x])&(not(is_this_salary(my_table_data1_2[x]))):
      city.append(my_table_data1_2[x])
      if (is_this_salary(my_table_data1_2[x+1])): 
        pass
      elif 'tahun' in my_table_data1_2[x+1]:
        salary.append('')
      elif bool(my_table_data1_2[x+1])==False:
        pass
      else:
        salary.append('')
        years.append('')
    elif (is_this_salary(my_table_data1_2[x])):
      salary.append(my_table_data1_2[x])
      if 'tahun' in my_table_data1_2[x+1]:
        pass
      elif ('tahun' not in my_table_data1_2[x+1])&(not(is_this_salary(my_table_data1_2[x+1]))):
        years.append('')
      elif bool(my_table_data1_2[x+1])==False:
        pass
      else:
        city.append('')
        years.append('')
    elif 'tahun' in my_table_data1_2[x]:
      years.append(my_table_data1_2[x])
      if ('tahun' not in my_table_data1_2[x+1])&(not(is_this_salary(my_table_data1_2[x+1]))):
        pass
      elif (is_this_salary(my_table_data1_2[x+1])):
        city.append('')
      elif bool(my_table_data1_2[x+1])==False:
        pass
      else:
        city.append('')
        salary.append('')
except IndexError: pass #avoid index error

#taking 8000 data
df = pd.DataFrame()
df['Job_Title'] = my_table_data1_1[0:8000]
df['City'] = city[0:8000]
df['Yrs_of_Exp'] = years[0:8000]
df['Salary'] = salary[0:8000]
df

#exporting to csv file
df.to_csv('DataCollection_Glints.csv',index=False)
