==> 3. ProjekPrediksiGaji_DataCleaning.ipynb <==
#3. Salary Prediction Project - Data Cleaning
by: Cindy Suyitno

Files that are used in this part:
- DataCollection_Glints.csv (from part 1. Data Collection)
- DataCollection_Karir.csv (from part 2. Data Collection)

File that is extracted from this part:
- saved_province.pkl (--> will be used in part 5. Webpage with Streamlit)
- DataCleaned.csv (--> will be used in part 4. Data Modelling)

#libraries that is not in google collab, need to be installed first
!pip install CurrencyConverter

#libraries used in this part
import pandas as pd
import numpy as np
from currency_converter import CurrencyConverter
from datetime import date

#importing all dataframes that we are going to clean
df_1 = pd.read_csv('DataCollection_Glints.csv')
df_2 = pd.read_csv('DataCollection_Karir.csv')
df_karir = pd.concat([df_2['Job_Title'],df_2['City'],df_2['Salary'],df_2['Yrs_of_Exp']], axis = 1) #only getting the same column as Data Collection from Glints
df_total = pd.concat([df_karir,df_1],axis = 0) 
df_total = df_total.reset_index().drop('index',1)
df_total = df_total.dropna(axis=0) #dropping any nan value because all of the columns will be used as predictor in data modelling
df_total.reset_index(drop=True, inplace=True)

df_total

#cleaning job title column
df_total['Job'] = df_total['Job_Title'].copy()

#assigning each jobs to job categories with a list of keywords
def replace_df_job(keyword, old_job, new_job):
  for key in keyword:
    if key in str.lower(old_job):
      df_total['Job'] = df_total['Job'].replace(old_job, new_job)    
  return df_total['Job']

a = ['akun','accou','finance','tax','pajak','fat','pay','cost','uang']
b = ['sales']
c = ['program']
d = ['product manager','produk manager']
e = ['factory','pabrik','operation','risk','supply','produksi','production','plan','qa','quality','kualitas','qc','ppic','inventory control','hse']
f = ['fashion','anima','design','desain','art','visual','ilustrator','illustrator','gambar']
g = ['purchas','beli']
h = ['developer']
i = ['analis','analyst']
j = ['admin','secret','sekret','asist','assist','document','dokumen','data entry']
k = ['teller','teler','bank']
l = ['model','advert','ads','creator','writer','creative','media','tiktok','tik','marketplace','market place','commerce','content','konten','digital','brand','copywriter']
m = ['electri','elektri']
n = ['hr','human resource','humas','rekrut','recruit','key performance']
o = ['law','legal','permit']
p = ['research','riset','scrum','specialist','spesialis']
q = ['guru','teach','tutor','dosen','learn','ajar']
r = ['event','bazaar','wedding']
s = ['supir','deliver','stok','stock','distribu','transpor','impor','pack','logis','warehouse','driver','kurir','ekspor','expor','truck','truk']
t = ['customer','call','costumer']
u = ['recept','receps','resep','cafe','kafe','barist','waiter','waitress','pramusaji','chef','cook','masak','kasir','cashier','hotel','resto','restaurant','bar','travel','tour','tur','ticket','tiket']
v = ['beauty','styl','spa','salon','nail','hair','make','rias','rambut']
w = ['health','sehat','dokter','medis','medic','farma','pharma','apotek','apotik','bidan']
x = ['video','photo','foto','kamera','camera']
y = ['ui','seo','software','frontend','front end','backend','back end','full','devops','web','chrome','cloud','network','iot','web']
z = ['engineer','maintenance','teknik','teknisi','mesin','tekhni','civil',' industrial','chemical','sipil','mekanik']
aa = ['data scien']
bb = ['proyek','arsi','archi','draft','constru','konstruk','site','surveyor','lapangan']
cc = ['public','relation','community']
dd = ['collect','kolek','asuransi','assurance']
ee = ['aman','safe','secur','sekur','satpam']
ff = ['clean','office boy','office girl','bersih','bantu']
gg = ['ikan','ternak','tani','kebun','agri','agro','farm']
hh = ['konsul','consul','qmr']
ii = ['translat','penerjemah']
jj = ['audit']
kk = ['Project Manager']
ll = ['kol']
mm = ['retail','merchan']
nn = ['jurnal','journal','report','editor','berita']
oo = ['entertain','talent']
pp = ['survey','survei']
qq = ['bisnis','business','owner','franchise','cabang','branch','manage','store','toko']
rr = ['it ','tech']
ss = ['market']
tt = ['management trainee','trainee']
uu = ['magang','intern ']

dictionary = [(a,'Accounting & Finance'),(b,'Sales'),(c,'Programmer'),(d,'Product Manager'),(e,'Factory and Manufacture'),(f,'Art & Design'),
                (g,'Purchasing'),(h,'Developer'),(i,'Analyst'),(j,'Administration'),(k,'Banking'),(l,'Internet & New Media'),(m,'Electrical'),
                (n,'Human Resources & General Affair'),(o,'Law'),(p,'Research & Development'),(q,'Teaching'),(r,'Event'),(s,'Transportation & Logistic'),
                (t,'Customer Service'),(u,'Hospitality & Travel'),(v,'Beauty'),(w,'Health'),(x,'Photography'),(y,'Software'),(z,'Engineering'),
                (aa,'Data Science'),(bb,'Construction and Building'),(cc,'Public Relation'),(dd,'Asuransi'),(ee,'Security'),(ff,'Cleaning'),(gg,'Agriculture'),
                (hh,'Consultant'),(ii,'Translator'),(jj,'Auditor'),(kk,'KOL'),(ll,'Retail'),(mm,'Journalistic'),(oo,'Entertainment'),(pp,'Survey Respondent'),
                (qq,'Business'),(rr,'General IT'),(ss,'Marketing'),(tt,'Trainee & Management Trainee'),(uu,'Internship')]
Job_list = []
for key,value in dictionary:
    Job_list.append(value)

for job in df_total['Job']:
  for key,value in dictionary:
    replace_df_job(key,str(job),value)
  if job not in Job_list:
    df_total['Job'] = df_total['Job'].replace(job, 'Other')

df_total["Job"].value_counts()

#recleaning due to minimum data and too many categories
a = ['product manager','retail','business']
b = ['teaching','translator']
c = ['beauty','event','entertainment','photography','journalistic']
d = ['electric','factory','trainee','internship','transportation','construction','engineering','purchasing']
e = ['analyst','data']
f = ['software','general it','developer','programmer']
g = ['agriculture','cleaning']
h = ['finance','banking','auditor']
i = ['customer','public']

dictionary = [(a,'Business'),(b,'Education'),(c,'Journalistic and Entertainment'),(d, 'Factory and Manufacturer'),(e, 'Data Analyst & Data Science'),
              (f,'Software Engineering'),(g,'Other'),(h,'Accounting, Finance, Banking'),(i,'Public Relation')]

for job in df_total['Job']:
  for key,value in dictionary:
    replace_df_job(key,str(job),value)

#final results for job column
df_total['Job'].value_counts()

#cleaning city column
df_total['Province'] = df_total['City'].copy()

#again, assigning each of the province with lists of keywords that can connect the city to the province
def replace_df_city(keyword, city, province):
  for key in keyword:
    if key in str.lower(city):
      df_total['Province'] = df_total['Province'].replace(city, province)    
  return df_total['Province']

for province in df_total['Province']:
  a = ['banten','ciputat','tangerang','cilegon','serang','pamulang','serpong','pagedangan','cikupa']
  b = ['jombang','madura','jawa timur','pamekasan','ponorogo','kalibuntu','tuban','situbondo', 'jember','banyuwangi','pasuruan','blitar','mojokerto','madiun','kediri','gresik','sidoarjo','bojonegoro','nganjuk','batu','surabaya','probolinggo','trenggalek','sidoaji','lamongan','malang']
  c = ['surakarta','jawa tengah','kebumen','purwodadi','kendal','pemalang','boyolali','grobogan','banjarnegara','salatiga','banyumas','pati','sragen','demak','batang','temanggung','kudus','karanganyar','sukoharjo','cilacap','jepara','klaten','magelang','solo','tegal','purwokerto','pekalongan','bandungan timur','brebes','semarang']
  d = ['jawa barat','cileungsi','garut','indramayu','sumedang','purwokerto','cibinong','kuningan','cibitung','kertasari','purwakarta','cimahi','tasikmalaya','grogol','sukabumi','karawang','cirebon','cikarang','cianjur','bandung','soreang','cikampek','ciamis','subang','sentul','bekasi','bogor','depok']
  e = ['sorong','manokwari']
  f = ['sumatera utara','siantar','sidimpuan','kisaran','tebing tinggi','balige','binjai','serdang','prapat','karo','adil makmur','medan','sibolga']
  g = ['senayan','kembangan','cilandak','kemang','abang','ancol','jakarta']
  h = ['tabanan','ubud','kuta','gianyar','asem','seminyak','denpasar','badung','singaraja','bali']
  i = ['banteng','makassar','sulawesi selatan','toraja','pare','palopo','bone']
  j = ['kalimantan timur','balikpapan','penajam','samarinda','bontang','sangatta','tarakan']
  k = ['pontianak','ketapang','sanggau']
  l = ['tanjungpinang','tanjung pinang','kepulauan']
  m = ['banjar baru','banjarmasin','pelaihari']
  n = ['padang','bukittinggi','solok','sumatera barat','pasaman']
  o = ['yogyakarta','sleman','bantul']
  p = ['bungku','nambo','palu','sulawesi tengah','sigi','poso']
  q = ['linggau','lahat','palembang','enim','prabumulih','sumatera selatan']
  r = ['manado','bitung','kotamobagu','sulawesi utara']
  s = ['parigi','kendari','konawe']
  t = ['batam','pekanbaru','hilir','dumai','pelalawan','riau']
  u = ['greater napanee','luar']
  v = ['kalimantan tengah','palangkaraya']
  w = ['bima','mataram','nusa tenggara barat']
  x = ['tanjung pandan','pangkal pinang']
  y = ['jambi','bungo']
  z = ['subulussalam','langsa','aceh','bireuen','meulaboh','lhokseumawe']
  aa = ['lampung','kotabumi','metro']
  bb = ['kupang','maumere','waingapu','flores','atambua']
  cc = ['nunukan','tanjung selor']
  dd = ['nan']
  ee = ['jayapura','papua']
  ff = ['ambon']
  gg = ['ternate']
  hh = ['mamuju']
  ii = ['apa saja']
  jj = ['gorontalo']
  kk = ['bengkulu']

  Dictionary = [(a,'Banten'),(b,'Jawa Timur'),(c,'Jawa Tengah'),(d,'Jawa Barat'),(e,'Papua Barat'),(f,'Sumatera Utara'),(g,'Jakarta'),(h,'Bali'),
                (i,'Sulawesi Selatan'),(j,'Kalimantan Timur'),(k,'Kalimantan Barat'),(l,'Kepulauan Riau'),(m,'Kalimantan Selatan'),(n,'Sumatera Barat'),
                (o,'Yogyakarta'),(p,'Sulawesi Tengah'),(q,'Sumatera Selatan'),(r,'Sulawesi Utara'),(s,'Sulawesi Tenggara'),(t,'Riau'),(u,'Overseas'),
                (v,'Kalimantan Tengah'),(w,'NTB'),(x,'Bangka Belitung'),(y,'Jambi'),(z,'Aceh'),(aa,'Lampung'),(bb,'NTT'),(cc,'Kalimantan Utara'),(dd,np.nan),
                (ee,'Papua'),(ff,'Maluku'),(gg,'Maluku Utara'),(hh,'Sulawesi Barat'),(ii,'Indonesia'),(jj,'Gorontalo'),(kk,'Bengkulu')]
          
  for key,value in Dictionary:
    replace_df_city(key,str(province),value)

df_total["Province"].value_counts()

#saving for geopandas visualization
import pickle
data = {'job': df_total['Job'],'province': df_total["Province"]}
with open('saved_province.pkl','wb') as file:
  pickle.dump(data, file)

#recleaning due to minimal data and too many categories
def replace_df_province(keyword, old_province, new_province):
  for key in keyword:
    if key in str.lower(old_province):
      df_total['Province'] = df_total['Province'].replace(old_province, new_province)    
  return df_total['Province']

a = ['banten','jawa','yogya']
b = ['sumatera','lampung','jambi','gorontalo','bengkulu','bangka belitung','riau','aceh']
c = ['sulawesi']
d = ['bali','ntt','ntb']
e = ['kalimantan']
f = ['papua','maluku','other','overseas','indonesia']
g = ['jakarta']

Dictionary = [(a,'Jawa'),(b,'Sumatera'),(c,'Sulawesi'),(d,'Bali dan Nusa Tenggara'),(e,'Kalimantan'),(f,'Other'),(g,'Jakarta')]

Province_list = []
for key,value in Dictionary:
    Province_list.append(value)

for province in df_total['Province']:
  for key,value in Dictionary:
    replace_df_province(key,str(province),value)

#final results for province column
df_total['Province'].value_counts()

#cleaning years column
df_total['Yrs_of_Exp'].value_counts()

#assigning minimum and maximum value of years of experiences
min_years = []
max_years = []
y = []

for x in df_total['Yrs_of_Exp']:
  y.append(str(x).split())

for x in y:
  if (x[0].isdigit()):
    if len(x)==2:
      min_years.append(0)
      max_years.append(0)
    else:
      min_years.append(x[0])
      max_years.append(x[2])
  elif (x[0] == 'Kurang')|(x[0] == 'Setidaknya'):
    min_years.append(0)
    max_years.append(1)
  elif x[0] == 'Lebih':
    min_years.append(10)
    max_years.append(20)
  elif x[0] == 'nan':
    min_years.append('')
    max_years.append('')

#cleaning salary column
df_total['Salary'].value_counts()

#assigning minimum and maximum salary
df_total['Salary'] = df_total['Salary'].replace('Login to view estimated salary by Glints.',np.nan)
df_total['Salary'] = df_total['Salary'].replace('Gaji Kompetitif',np.nan)
df_total = df_total.dropna(axis=0)
df_total.reset_index(drop=True, inplace=True)

y = []
min_salary = []
max_salary = []
cc = CurrencyConverter()

for salary in df_total['Salary']:
  y.append(str(salary).strip().replace('.','').split())

for x in y:
   if x[0] == 'IDR':
     min_salary.append(x[1].strip(''))
     max_salary.append(x[-1])
   elif x[0] == '0':
     min_salary.append(0)
     max_salary.append(0)
   elif x[0] == 'USD':
     min_salary.append(cc.convert(x[1],'USD','IDR'))
     max_salary.append(cc.convert(x[-1],'USD','IDR'))
   elif x[0] == 'SGD':
     min_salary.append(cc.convert(x[1],'SGD','IDR'))
     max_salary.append(cc.convert(x[-1],'SGD','IDR'))
   elif x[0] == 'AUD':
     min_salary.append(cc.convert(x[1],'AUD','IDR'))
     max_salary.append(cc.convert(x[-1],'AUD','IDR'))
   elif x[0] == 'HKD':
     min_salary.append(cc.convert(x[1],'HKD','IDR'))
     max_salary.append(cc.convert(x[-1],'HKD','IDR'))

#join all dataframes together
df_cleaned_data = pd.concat([df_total['Job'],df_total['Province'],pd.DataFrame(min_years),pd.DataFrame(max_years),pd.DataFrame(min_salary),pd.DataFrame(max_salary)],axis=1)
df_cleaned_data.columns=['Job','Province','Min_Years','Max_Years','Min_Salary','Max_Salary']
df_cleaned_data

#export to csv
df_cleaned_data.to_csv('DataCleaned.csv',index=False)
